Session-class {SparkBeyond}	R Documentation

SB object that encapsulates a session

Description

SB object that encapsulates a session

Fields

artifact_loc

String location pointing to the model artifact.

modelBuilt

Indication for whether only a feature search was performed or a full model was created.

Methods

buildNumber()

Returns the build number in which the model was generated.

createPackage(sampleData = NULL, createRestAPIpackage = FALSE, ...)

Create a sharable package for the model. sampleData can be used to a sample data to the package and test it. Only first 20 rows of the sample data will be used. createRestAPIpackage is a boolean indicator for whether to create a package for prediction via command line (set to FALSE) or via programmatic REST API call(TRUE).

enrich(data, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. data is a dataframe to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

enrich.file(file, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. file is a path to the file to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

evaluate()

Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.

features()

Returns a dataset with top feature information

initialize(artifact_loc, modelBuilt = FALSE)

initializes a session using a string provided as loc.

liftFromPrediction(predictionResult, overrideDesiredClass = NA, title = NA, percentOfPopulationToPlot = 0.2)

Returns lift from a created model and generates three plots. predictionResult is a dataframe to be analyzed, overrideDesiredClass the class in the label column to check the lift for (e.g. '1'), title optional: a title for the plot. percentOfPopulationToPlot optional: limit the plot to the top percent of the data (x axis).

predict(data, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. data is a dataframe to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

predict.file(file, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. file is the path of the file to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

showConfusionMatrix(normalized = FALSE)

Shows a confusion matrix of a model.

showExtractors()

Shows extractors.

showFeatureClusters()

Shows the representative feature clusters pdf report.

showFeaturesTest()

Shows features performance on test.

showFeaturesTrain()

Shows features performance on train.

showFields()

Shows fields.

showFunctions()

Shows functions.

showInputSchema()

Shows the input schema.

showModelComparison()

Shows cross validation of various algorithms tested to create a model.

showROC()

Shows ROC of the model.

showROC_CV()

Shows ROC of cross validation of various algorithms tested to create a model.

showReport(report_name = NA)

report_name should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, train_features, train_unweighted_features, test_unweighted_features,field, function, InputSchema, modelComparison, roc_best, roc_CV

status()

Checking the status of the session.

waitForProcess()

Blocking the R console until session is finished.

Examples

# Learning example
## No test: 
session = learn("titanic", getData("titanic_train"), target="survived",algorithmsWhiteList = list("xgboostClassifier"),  scoreOnTestSet = TRUE, useCachedFeatures=TRUE)

## [1] "Notice: you are currently not using the latest engine version. Please consider running restartServer()."
## [1] "SparkBeyond package is outdated. Please consider updatePackage()"
## [1] "Calling: http://127.0.0.1:9000/rapi/learn"
## [1] "Training on  /tmp/titanic_train_ab99a20f8ea70e75b52c4ae61c0c3321.tsv"
## [1] "Artifact location was created at: /Users/zinman/localdata/artifacts/titanic/4"
## [1] "auto saved Session object to a variable named ' backup_titanic_4 '. To retrieve use: load('/Users/zinman/sparkBeyond/R/SparkBeyond/html/backup_titanic_4.RData')."
## [1] "Detecting types - 1"
## Feature Search training-set: 710.0
## Original histogram (FS):
## +----------+-----------+
## | Survived | Frequency |
## +----------+-----------+
## | 0        | 442       |
## | 1        | 268       |
## +----------+-----------+
## Prior entropy: 0.9562317858093743
## 
## Input Schema
## +-------+-----------+---------------+
## | index | name      | type          |
## +-------+-----------+---------------+
## | 0     | *survived | Int           |
## | 1     | pclass    | Int           |
## | 2     | name      | core.nlp.Text |
## | 3     | sex       | String        |
## | 4     | age       | Double        |
## | 5     | sibsp     | Int           |
## | 6     | parch     | Int           |
## | 7     | ticket    | core.nlp.Text |
## | 8     | fare      | Double        |
## | 9     | cabin     | String        |
## | 10    | embarked  | String        |
## +-------+-----------+---------------+
## [1] "Session in progress:  Creating features - 2"
## [1] "Printing top 50 features out of 300"
##    idx
## 1    0
## 2    1
## 3    2
## 4    3
## 5    4
## 6    5
## 7    6
## 8    7
## 9    8
## 10   9
## 11  10
## 12  11
## 13  12
## 14  13
## 15  14
## 16  15
## 17  16
## 18  17
## 19  18
## 20  19
## 21  20
## 22  21
## 23  22
## 24  23
## 25  24
## 26  25
## 27  26
## 28  27
## 29  28
## 30  29
## 31  30
## 32  31
## 33  32
## 34  33
## 35  34
## 36  35
## 37  36
## 38  37
## 39  38
## 40  39
## 41  40
## 42  41
## 43  42
## 44  43
## 45  44
## 46  45
## 47  46
## 48  47
## 49  48
## 50  49
##                                                                                                                 feature
## 1                                                                                              name does not contain mr
## 2                                                                                                           pclass != 3
## 3                                                                            numberIn(ticket) in range (8854.5, 231932)
## 4                                                                                            fare >= 15.172899999999998
## 5                                                                                numericSuffix(cabin) in range (3, 108)
## 6                                                                                                           sex != male
## 7                                                        cos(fare) in range (-0.4142747865977665, -0.18454433209625384)
## 8                                                                                                         embarked == C
## 9                                                                                                           pclass == 1
## 10                                                                                                      name contains )
## 11                                           upperCaseRatio(ticket) in range (0.05000000000000001, 0.13392857142857142)
## 12                                                                                      cos(fare) < -0.9989278625592983
## 13                                                                                                    round(fare) == 19
## 14                                                                                             numberOfWords(name) == 7
## 15                                                           sin(fare) in range (0.998425635754864, 0.9995472778889662)
## 16                                                           tan(age) in range (0.1652032253816651, 1.1682874182779286)
## 17                                                                                             numberOfWords(name) >= 6
## 18                                                                              smallSubsets(cabin) contains List(2, E)
## 19                                                                                                 sigmoid(fare) == 1.0
## 20                                                                                                           sibsp >= 5
## 21                                             wordFrequencies(name).valueOf(john) orIfMissing 0 >= 0.47360679774997894
## 22                                                             wordFrequencies(name) contains (mrs, 0.3779644730092272)
## 23                                                                                          prefixes(cabin) contains E1
## 24                                                                                        upperCaseRatio(name) == 0.125
## 25                                                                         normalizedNgrams(name) contains (mr, joseph)
## 26                                                                                                    ticket contains o
## 27                                                                                                   floor(fare) < 10.5
## 28                                                                           upperCaseRatio(name) < 0.15301003344481606
## 29                                                                                               numberIn(cabin) == 2.0
## 30                                                                                                    ceil(fare) == 7.0
## 31                                                                         normalizedNgrams(name) contains (miss, anna)
## 32                                                                                         numberIn(ticket) >= 319372.5
## 33                                                                              smallSubsets(cabin) contains List(5, 2)
## 34                                                                                      sin(age) < -0.21049876125431888
## 35                                                             stemmedWordsWithoutCommonTerms(name) does not contain mr
## 36                                                                                     Histogram(cabin) contains (2, 1)
## 37                                                                                                   ceil(fare) == 35.0
## 38                                                                                                        embarked != S
## 39                                                                                                          age == 47.0
## 40                                                                                  normalizedNgrams(name) contains (j)
## 41                                                                                    mostCommonValue(cabin) == Some(9)
## 42                                                                              smallSubsets(cabin) contains List(1, D)
## 43                                                                          sentimentDistribution(name) contains (0, 3)
## 44                                                                                                  name contains david
## 45                                                                                                   name contains kate
## 46                                         wordFrequencies(name).valueOf(margaret) orIfMissing 0 >= 0.47360679774997894
## 47                                            wordFrequencies(name).valueOf(alfred) orIfMissing 0 >= 0.4125890342545926
## 48                                                                                       upperCaseRatio(name) == 0.1667
## 49                                                              wordFrequencies(name) contains (mr, 0.5773502691896258)
## 50 wordFrequencies(name).valueOf(john) orIfMissing 0.5773502691896258 in range (0.3657589318012504, 0.4277309429819105)
##            RIG lin..score       support
## 1  0.246524231  0.4428309 299 (42.11%) 
## 2  0.076231905  0.4828444 315 (44.37%) 
## 3  0.071115835  0.6485446 199 (28.03%) 
## 4  0.068179011  0.4857528 333 (46.90%) 
## 5  0.065055070  0.6512270 145 (20.42%) 
## 6  0.242667723  0.4214526 251 (35.35%) 
## 7  0.026757770  0.6091381  72 (10.14%) 
## 8  0.023689570  0.5007106 139 (19.58%) 
## 9  0.062929319  0.4867891 172 (24.23%) 
## 10 0.010550407  0.5306137  14 (1.97%)  
## 11 0.018753223  0.6129032  26 (3.66%)  
## 12 0.012511649  0.5306704   6 (0.85%)  
## 13 0.012511649  0.5306704   6 (0.85%)  
## 14 0.032384616  0.5229437  29 (4.08%)  
## 15 0.023655675  0.6114767  31 (4.37%)  
## 16 0.010770463  0.6141236 104 (14.65%) 
## 17 0.058777034  0.5127108  82 (11.55%) 
## 18 0.012511649  0.5306704   6 (0.85%)  
## 19 0.046715422  0.5171947 163 (22.96%) 
## 20 0.011189840  0.5065561  11 (1.55%)  
## 21 0.014925332  0.6140232  22 (3.10%)  
## 22 0.034832967  0.5228582  21 (2.96%)  
## 23 0.010413798  0.5313050   5 (0.70%)  
## 24 0.013092465  0.5296784  18 (2.54%)  
## 25 0.008115772  0.5325901   8 (1.13%)  
## 26 0.009393665  0.5318326  16 (2.25%)  
## 27 0.064727877  0.5118701 299 (42.11%) 
## 28 0.044987230  0.5183965 337 (47.46%) 
## 29 0.011035286  0.5849946   9 (1.27%)  
## 30 0.010163270  0.5321201  10 (1.41%)  
## 31 0.016722656  0.5293936   8 (1.13%)  
## 32 0.027003283  0.5796781  99 (13.94%) 
## 33 0.010413798  0.5313050   5 (0.70%)  
## 34 0.011542697  0.5302979 248 (34.93%) 
## 35 0.093387963  0.4997182 197 (27.75%) 
## 36 0.019477192  0.5272582  38 (5.35%)  
## 37 0.006075752  0.5330571   6 (0.85%)  
## 38 0.026285346  0.4998902 208 (29.30%) 
## 39 0.007094831  0.5074514   7 (0.99%)  
## 40 0.007094831  0.5328240   7 (0.99%)  
## 41 0.010413798  0.5313050   5 (0.70%)  
## 42 0.010413798  0.5313050   5 (0.70%)  
## 43 0.017400935  0.5283457 279 (39.30%) 
## 44 0.006075752  0.5330571   6 (0.85%)  
## 45 0.012511649  0.5306704   6 (0.85%)  
## 46 0.012511649  0.6143173   6 (0.85%)  
## 47 0.009138583  0.6162681   9 (1.27%)  
## 48 0.015498520  0.5294405  72 (10.14%) 
## 49 0.090437205  0.5055971 182 (25.63%) 
## 50 0.012511649  0.7111491   6 (0.85%)  
## [1] "Session in progress:  Building model - 3"
## [1] "Session in progress:  Building model - 4"
## [1] "Session in progress:  Building model - 5"
## Score (AUC): 0.85539
## Previous (revision 3) Score (AUC): 0.85116
##                                                
## Correctly Classified Instances     148   81.77%
## Incorrectly Classified Instances   33    18.23%
## Total Number of Instances          181         
##                                                
## === Detailed Accuracy By Class ===
##                                                                                      
## Class           TP Rate   FP Rate   Precision   Recall   F-Measure   ROC Area   Gini 
##                                                                                      
## 0               0.944     0.365     0.789       0.944    0.86        0.855      0.711
## 1               0.635     0.056     0.887       0.635    0.74        0.855      0.711
## Weighted Avg.   0.818     0.239     0.829       0.818    0.811       0.855      0.711
##                                                                                      
## Confusion Matrix:
## +---------------------+-----+----+
## | Actual \ Prediction | 0   | 1  |
## +---------------------+-----+----+
## | 0                   | 101 | 6  |
## | 1                   | 27  | 47 |
## +---------------------+-----+----+
## [1] "Done"

#session = featureSearch("titanic", getData("titanic_train"), target="survived")
enriched = session$enrich(getData("titanic_train"), featureCount = 10)

## [1] "Notice: you are currently not using the latest engine version. Please consider running restartServer()."
## [1] "Enriching  /tmp/enrich_ab99a20f8ea70e75b52c4ae61c0c3321.tsv"
## [1] "Calling: http://127.0.0.1:9000/rapi/enrich"
## [1] "Done."

colnames(enriched)

##  [1] "X.name.contains.mr"                                        
##  [2] "pclass....3"                                               
##  [3] "abs.numberIn.ticket....120393.25....111538.75"             
##  [4] "fare....15.172899999999998"                                
##  [5] "abs.numericSuffix.cabin....55.5....52.5"                   
##  [6] "sex....male"                                               
##  [7] "abs.cos.fare.....0.2994095593470102....0.11486522725075635"
##  [8] "embarked....C"                                             
##  [9] "pclass....1"                                               
## [10] "name.contains.."

predicted = session$predict(getData("titanic_test"))

## [1] "Notice: you are currently not using the latest engine version. Please consider running restartServer()."
## [1] "Predicting  /tmp/predict_0f89734c38a6ae1ded7bba3717abc94f.tsv"
## [1] "Calling: http://127.0.0.1:9000/rapi/predict"
## [1] "Predictions and plots are available at: /Users/zinman/localdata/artifacts/titanic/4/reports/predictions/test/"
## [1] "Done."

colnames(predicted)

## [1] "survived_predicted" "probability_0"      "probability_1"

predicted[1:5,c("survived_predicted", "probability_0", "probability_1")]

##   survived_predicted probability_0 probability_1
## 1                  0     0.9209693    0.07903069
## 2                  1     0.3959064    0.60409361
## 3                  0     0.8407071    0.15929288
## 4                  0     0.6627080    0.33729205
## 5                  1     0.3373680    0.66263205

eval = session$evaluate()

## === Detailed Accuracy By Class ===
##                                                                                      
## Class           TP Rate   FP Rate   Precision   Recall   F-Measure   ROC Area   Gini 
##                                                                                      
## 0               0.944     0.365     0.789       0.944    0.86        0.855      0.711
## 1               0.635     0.056     0.887       0.635    0.74        0.855      0.711
## Weighted Avg.   0.818     0.239     0.829       0.818    0.811       0.855      0.711
## 

#session$showFeatures()
#session$showConfusionMatrix()
## End(No test)

[Package SparkBeyond version 0.1 Index]