colSizesUnique {SparkBeyond}	R Documentation

colSizesUnique

Description

Count the unique number of items in each column

Usage

colSizesUnique(data)

Arguments

data
dataframe / data.table to be examined

Value

a summary of unique items counts per row

Examples

grouped = groupBy(getData("titanic_train"), by ="pclass")
rownames(grouped)

## [1] "1" "2" "3"

colSizes(grouped)

##   pclass survived name sex age sibsp parch ticket fare cabin embarked
## 1      1      491  491 491 491   491   491    491  491   491      491
## 2      1      216  216 216 216   216   216    216  216   216      216
## 3      1      184  184 184 184   184   184    184  184   184      184

colSizesUnique(grouped)

##   pclass survived name sex age sibsp parch ticket fare cabin embarked
## 1      1        2  491   2  69     7     7    394  119     8        3
## 2      1        2  216   2  58     4     4    147   94   134        4
## 3      1        2  184   2  58     4     4    140   42     8        3

head(grouped)

##    pclass     survived
## 1:      3 0,1,0,0,1,0,
## 2:      1 1,1,0,1,1,0,
## 3:      2 1,1,1,0,1,0,
##                                                                                                                                                                                                              name
## 1:                                     Braund, Mr. Owen Harris,Heikkinen, Miss. Laina,Allen, Mr. William Henry,Palsson, Master. Gosta Leonard,Johnson, Mrs. Oscar W (Elisabeth Vilhelmina Berg),Moran, Mr. James,
## 2: Cumings, Mrs. John Bradley (Florence Briggs Thayer),Futrelle, Mrs. Jacques Heath (Lily May Peel),McCarthy, Mr. Timothy J,Bonnell, Miss. Elizabeth,Sloper, Mr. William Thompson,Fortune, Mr. Charles Alexander,
## 3:                                            Nasser, Mrs. Nicholas (Adele Achem),Hewlett, Mrs. (Mary D Kingcome) ,Williams, Mr. Charles Eugene,Fynney, Mr. Joseph J,Beesley, Mr. Lawrence,Wheadon, Mr. Edward H,
##                                     sex                age        sibsp
## 1:   male,female,male,male,female,male, 22,26,35, 2,27,NA, 1,0,0,3,0,0,
## 2: female,female,male,female,male,male, 38,35,54,58,28,19, 1,1,0,0,0,3,
## 3:   female,female,male,male,male,male, 14,55,NA,35,34,66, 1,0,0,0,0,0,
##           parch                                                  ticket
## 1: 0,0,0,1,2,0, A/5 21171,STON/O2. 3101282,373450,349909,347742,330877,
## 2: 0,0,0,0,0,2,              PC 17599,113803,17463,113783,113788,19950,
## 3: 0,0,0,0,0,0,          237736,248706,244373,239865,248698,C.A. 24579,
##                                                      fare
## 1:        7.2500, 7.9250, 8.0500,21.0750,11.1333, 8.4583,
## 2:  71.2833, 53.1000, 51.8625, 26.5500, 35.5000,263.0000,
## 3:       30.0708,16.0000,13.0000,26.0000,13.0000,10.5000,
##                                cabin     embarked
## 1:                            ,,,,,, S,S,S,S,S,Q,
## 2: C85,C123,E46,C103,A6,C23 C25 C27, C,S,S,S,S,S,
## 3:                         ,,,,D56,, C,S,S,S,S,S,

grouped2 = groupBy(getData("titanic_train"), by =list("pclass","sex"))
rownames(grouped2)

## [1] "1" "2" "3" "4" "5" "6"

colSizes(grouped2)

##   pclass sex survived name age sibsp parch ticket fare cabin embarked
## 1      1   1      347  347 347   347   347    347  347   347      347
## 2      1   1       94   94  94    94    94     94   94    94       94
## 3      1   1      144  144 144   144   144    144  144   144      144
## 4      1   1      122  122 122   122   122    122  122   122      122
## 5      1   1       76   76  76    76    76     76   76    76       76
## 6      1   1      108  108 108   108   108    108  108   108      108

colSizesUnique(grouped2)

##   pclass sex survived name age sibsp parch ticket fare cabin embarked
## 1      1   1        2  347  64     7     6    311   98     6        3
## 2      1   1        2   94  42     4     3     66   61    68        4
## 3      1   1        2  144  47     7     7    115   69     4        3
## 4      1   1        2  122  50     4     4    114   70    88        3
## 5      1   1        2   76  41     4     4     66   28     6        3
## 6      1   1        2  108  45     3     3     94   31     5        3

head(grouped2)

##    pclass    sex     survived
## 1:      3   male 0,0,0,0,0,0,
## 2:      1 female 1,1,1,1,1,1,
## 3:      3 female 1,1,1,0,0,1,
## 4:      1   male 0,1,0,0,0,0,
## 5:      2 female 1,1,0,1,1,1,
## 6:      2   male 1,0,1,0,0,0,
##                                                                                                                                                                                                                                            name
## 1:                                                                                 Braund, Mr. Owen Harris,Allen, Mr. William Henry,Palsson, Master. Gosta Leonard,Moran, Mr. James,Saundercock, Mr. William Henry,Andersson, Mr. Anders Johan,
## 2:       Cumings, Mrs. John Bradley (Florence Briggs Thayer),Futrelle, Mrs. Jacques Heath (Lily May Peel),Bonnell, Miss. Elizabeth,Spencer, Mrs. William Augustus (Marie Eugenie),Harper, Mrs. Henry Sleeper (Myna Haxtun),Icard, Miss. Amelie,
## 3:               Heikkinen, Miss. Laina,Johnson, Mrs. Oscar W (Elisabeth Vilhelmina Berg),Sandstrom, Miss. Marguerite Rut,Vestrom, Miss. Hulda Amanda Adolfina,Vander Planke, Mrs. Julius (Emelia Maria Vandemoortele),Masselmani, Mrs. Fatima,
## 4:                                                                         McCarthy, Mr. Timothy J,Sloper, Mr. William Thompson,Fortune, Mr. Charles Alexander,Uruchurtu, Don. Manuel E,Meyer, Mr. Edgar Joseph,Holverson, Mr. Alexander Oskar,
## 5: Nasser, Mrs. Nicholas (Adele Achem),Hewlett, Mrs. (Mary D Kingcome) ,Turpin, Mrs. William John Robert (Dorothy Ann Wonnacott),Laroche, Miss. Simonne Marie Anne Andree,Faunthorpe, Mrs. Lizzie (Elizabeth Anne Wilkinson),Rugg, Miss. Emily,
## 6:                                                                                               Williams, Mr. Charles Eugene,Fynney, Mr. Joseph J,Beesley, Mr. Lawrence,Wheadon, Mr. Edward H,Jenkin, Mr. Stephen Curnow,Hood, Mr. Ambrose Jr,
##                   age        sibsp        parch
## 1: 22,35, 2,NA,20,39, 1,0,3,0,0,1, 0,0,1,0,0,5,
## 2: 38,35,58,NA,49,38, 1,1,0,1,1,0, 0,0,0,0,0,0,
## 3: 26,27, 4,14,31,NA, 0,0,1,0,1,0, 0,2,1,0,0,0,
## 4: 54,28,19,40,28,42, 0,0,3,0,1,1, 0,0,2,0,0,0,
## 5: 14,55,27, 3,29,21, 1,0,1,1,1,0, 0,0,0,2,0,0,
## 6: NA,35,34,66,32,21, 0,0,0,0,0,0, 0,0,0,0,0,0,
##                                                      ticket
## 1:         A/5 21171,373450,349909,330877,A/5. 2151,347082,
## 2:         PC 17599,113803,113783,PC 17569,PC 17572,113572,
## 3:      STON/O2. 3101282,347742,PP 9549,350406,345763,2649,
## 4:             17463,113788,19950,PC 17601,PC 17604,113789,
## 5:       237736,248706,11668,SC/Paris 2123,2926,C.A. 31026,
## 6: 244373,239865,248698,C.A. 24579,C.A. 33111,S.O.C. 14879,
##                                                      fare
## 1:        7.2500, 8.0500,21.0750, 8.4583, 8.0500,31.2750,
## 2:  71.2833, 53.1000, 26.5500,146.5208, 76.7292, 80.0000,
## 3:        7.9250,11.1333,16.7000, 7.8542,18.0000, 7.2250,
## 4:  51.8625, 35.5000,263.0000, 27.7208, 82.1708, 52.0000,
## 5:       30.0708,16.0000,21.0000,41.5792,26.0000,10.5000,
## 6:                         13.0,26.0,13.0,10.5,10.5,73.5,
##                         cabin     embarked
## 1:                     ,,,,,, S,S,S,Q,S,S,
## 2: C85,C123,C103,B78,D33,B28,  C,S,S,C,C,,
## 3:                   ,,G6,,,, S,S,S,S,S,C,
## 4:     E46,A6,C23 C25 C27,,,, S,S,S,C,C,S,
## 5:                     ,,,,,, C,S,S,C,S,S,
## 6:                  ,,D56,,,, S,S,S,S,S,S,

[Package SparkBeyond version 0.1 Index]