contextObject {SparkBeyond}	R Documentation

contextObject

Description

contextObject

Usage

contextObject(data, name = NULL, keyColumns = list(), timeColumn = NULL)

Arguments

data:
a data frame or a path to a file containing the context data

name:
an identifier for the context object to be created (optional).

keyColumns:
Specify the key columns to be used by the context object (optional).

timeColumn:
Specify the time column to be used by the context object (relevant in time series contexts) (optional).

[Package SparkBeyond version 0.1 Index]