distinctBy {SparkBeyond}	R Documentation

distinctBy

Description

Keep only distinct rows in the input as specified by parameter "by". When multiple rows are encountered for a certain combination the first row will be selected.

Usage

distinctBy(data, by)

Arguments

data:
dataframe or data.table to be grouped

by:
a list of column names to distinct by

Value

a new data.table object with the distinct data

[Package SparkBeyond version 0.1 Index]