excludeCols {SparkBeyond}	R Documentation

excludeCols

Description

Exclude columns from a data frame. Please notice that the function return value depends on the type of input. NA will be returned if the input is of type data.table and the input object will be modified. Otherwise a data.frame object will be returned.

Usage

excludeCols(data, cols)

Arguments

data:
dataframe / data data.table to modify.

cols:
a list of column names to remove.

Value

if the input data is a dataframe than a dataframe with the excluded columns will be return. If the input data is a data.table object, NA will be returned and the input object will be modified.

Examples

data = getData("titanic_train")

## data.table case
datatable = data.table(data)
class(datatable)

## [1] "data.table" "data.frame"

colnames(datatable)

##  [1] "survived" "pclass"   "name"     "sex"      "age"      "sibsp"   
##  [7] "parch"    "ticket"   "fare"     "cabin"    "embarked"

excludeCols(datatable, list("sex", "age"))

## [1] "sex, age were removed"
## [1] "Note: the input object is of type data.table hence NA is returned and the input object will be modified,"

## [1] NA

colnames(datatable)

## [1] "survived" "pclass"   "name"     "sibsp"    "parch"    "ticket"  
## [7] "fare"     "cabin"    "embarked"

## dataframe case
class(data)

## [1] "data.frame"

colnames(data)

##  [1] "survived" "pclass"   "name"     "sex"      "age"      "sibsp"   
##  [7] "parch"    "ticket"   "fare"     "cabin"    "embarked"

excluded = excludeCols(data, list("sex", "age"))

## [1] "sex, age were removed"
## [1] "Note: the input object is of type data.frame hence, a new data.frame with the excluded columns will be returned."

colnames(excluded)

## [1] "survived" "pclass"   "name"     "sibsp"    "parch"    "ticket"  
## [7] "fare"     "cabin"    "embarked"

[Package SparkBeyond version 0.1 Index]