featureGenerationControl {SparkBeyond}	R Documentation

featureGenerationContorl

Description

featureGenerationContorl

Usage

featureGenerationControl(maxDepth = 2, featureSearchMode = "DEFAULT",
  functionsWhiteList = NA, functionsBlackList = NA,
  booleanNumericFeatures = NA, numericEqualityFeatures = TRUE,
  allowRangeFeatures = TRUE, crossRowFeatureSearch = FALSE,
  maxFeaturesCount = list(300), autoColumnSubSets = list("CONCEPT"),
  customColumnSubsets = NA, maxFeatureDuration = 100,
  overrideMaxFeatureDurationForExternalData = TRUE,
  allocatedMemoryMB = 1000, maxCollectionSize = NA,
  useCachedFeatures = TRUE)

Arguments

maxDepth:
Optional. Integer < 8 which represent the maximum number of transformations allowed during the feature search phase. Increasing this value should be considered with cautious as the feature search phase is exponential. 2 by default.

featureSearchMode:
One of the following DEFAULT, ADVANCED, ADVANCED_WITH_PAIRS, DIG_DEEP

functionsWhiteList:
Optional. A list of strings that represents a set of functions that will be used to guide the feature search. NA by default.

functionsBlackList:
Optional. A list of strings that represents a set of function that will be excluded from the feature search. Can also include function domains including('math','arithmetics', 'collections', 'booleanOperators', 'semantics', 'nlp', 'trigonometry', 'bitwise'). NA by default.

booleanNumericFeatures:
A boolean indicating whether to transform all features to boolean values. (i.e., when FALSE the continuous value of the feature left-hand-side will be passed to the algorithm, without taking into account the specific cutoff chosen during the feature search phase). NA by default indicating that it will be TRUE for classification problems and FALSE for regression problems.

numericEqualityFeatures:
A boolean indicator for whether to include features that compare numeric fields with specific values. TRUE by default.

allowRangeFeatures:
A boolean indicator for whether to include features that define range over a set of numeric values. TRUE by default.

crossRowFeatureSearch.
A booleean indicating whether to allow creating features using data collected from multiple rows together.FALSE by default.

maxFeaturesCount:
Optional. A list of integers indicating how many features should be created by the SB engine. 300 by default.

autoColumnSubSets:
Optional. A list of values contain any of the following: "CONCEPT", "NUMERIC_PAIRS", "ALL_PAIRS". "CONCEPT" will aim to generate column subset from fields that are from similar non-numeric types or combination of date and non-numeric elements. "NUMERIC_PAIRS" will create all column subsets for numeric columns. "ALL_PAIRS" will create subsets of size 2 from all columns. "CONCEPT" by default.

customColumnSubsets:
Optional. A List of lists containing specific column subsets to examine. In order to set a certain depth to the subset, add an element in the end of the customSubSet with one digit as a string representing the requested depth. If not such element was defined the regular depth definitions will be used. NA by default.

maxFeatureDuration:
Optional. A numeric value representing the maximum allowed time a feature may take during search per row in milliseconds. 100 by default.

overrideMaxFeatureDurationForExternalData:
A boolean indicating whether the maxFeatureDuration parameter should not be used for features that use external data. TRUE by default.

allocatedMemoryMB:
Optional. Integer value representing how to chunk the memory during feature search . 1000MB by default.

maxCollectionSize:
Optional. Integer value repsenting what is the maximum cardinality allowed for a transformation during feature search. NA by default corresponding to 200K.

useCachedFeatures:
Optional. A boolean indicating whether to use cached features (from previous run). TRUE by default.

[Package SparkBeyond version 0.1 Index]