isLatestVersion {SparkBeyond}	R Documentation

A function to verify if we are using the latest server version

Description

A function to verify if we are using the latest server version

Usage

isLatestVersion()

Value

Boolean indicating TRUE if we are using the latest version otherwise FALSE.

[Package SparkBeyond version 0.1 Index]