isServerAlive {SparkBeyond}	R Documentation

A function to check whether the server is alive

Description

A function to check whether the server is alive

Usage

isServerAlive()

Value

The response from the server.

[Package SparkBeyond version 0.1 Index]