knowledgeControl {SparkBeyond}	R Documentation

knowledge Control

Description

knowledge Control

Usage

knowledgeControl(useGraph = FALSE, useOpenStreetMap = FALSE,
  useCustomGraphs = FALSE, customGraphsWhiteList = NA,
  customGraphsBlackList = NA, customFunctions = NA)

Arguments

useGraph:
Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.

useOpenStreetMap:
Optional. A boolean indicating whether Open Street Map data should be included in feature search. FALSE by default.

useCustomGraphs:
A boolean indicating whether custom graphs should be used. FALSE by default.

customGraphsWhiteList:
Optional. A list that filters which domains should be used. NA by default.

customGraphsBlackList:
Optional. A list that filters which custom graphs should be ignored. NA by default.

customFunctions:
Optional. A list of additional functions that should be incorporated to the feature search. NA by default.

[Package SparkBeyond version 0.1 Index]