learn {SparkBeyond}	R Documentation

learn

Description

Runs SparkBeyond feature enrichment and learning process.

Usage

learn(projectName = "temp", trainData, target, testData = NA,
  trainTestSplitRatio = 0.8, weightColumn = NA, weightByClass = FALSE,
  contextDatasets = NA, preProcessingCtrl = preProcessingControl(),
  featureGenerationCtrl = featureGenerationControl(),
  knowledgeCtrl = knowledgeControl(),
  modelBuildingCtrl = modelBuildingControl(),
  reportingCtrl = reportingControl(), verbose = FALSE, autoSave = TRUE,
  runBlocking = TRUE, ...)

Arguments

target
String of the column name of in the training file that contains the target of the prediction.

projectName:
Optional string of the session name. Setting a session name is highly recommended. "temp" by default.

trainData:
train data to analyze.

testData:
Optional. test data to validate model results. NA by default.

trainTestSplitRatio:
Optional. Double value in [0,1] to split the train file data in order to keep some data for test. 0.8 by default. Ignored if test filename was provided.

weightColumn:
Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.

weightByClass:
Adds a weight column with values inverse proportional to the frequency of the class. FALSE by default.

contextDatasets:
Optional. A list of paths to context datasets to be added to the learning.

featureGenerationControl:
A featureGenerationControl object with specific feature generation parameters.

knowledgeControl:
A knowledgeControl object with specific external knowledge parameters.

modelBuildingControl:
A modelBuildingControl object with specific model building parameters.

reportingControl:
A reportingControl object with specific reporting parameters.

autoSave:
Optional. Automatically saves the generated session object to a file for future use. Good in cases where the connection between R and the server was interrupted or you would like to review previous models results. TRUE by default.

runBlocking:
Block the R console while the session is running. FALSE by default.

Value

Session object that encapsulates the model.

Examples

#session = learn("titanic", getData("titanic_train"), target = "survived", algorithmsWhiteList = list("RRandomForest"), runBlocking = TRUE, autoSave=FALSE)

[Package SparkBeyond version 0.1 Index]