limitTimeSeries {SparkBeyond}	R Documentation

limitTimeSeries

Description

filter rows to contain only rows that are in a certain time frame based on a from/until date inputs (inclusive). All dates are assumed to be of the same format. This function should be called before any filtering is performed.

Usage

limitTimeSeries(data, dateCol = "SB_times_col", fromDate = NA,
  untilDate = NA, datesFormat = "%m/%d/%Y")

Arguments

data:
the dataframe / data.table to be modified.

dateCol:
The column name in data that will be used for filtering. "SB_times_col" by default

fromDate:
The starting date to filter from. NA by default.

untilData:
The end date to filter until. NA by default.

datesFormat:
the format of the from/until dates.month/day/year by default.

Value

a new data.table with the filtered rows will be returned.

Examples

randDate <- function(N, st="2014/01/01", et="2014/12/31") {
 st <- as.POSIXct(as.Date(st,tz = "EST"),tz = "EST")
 et <- as.POSIXct(as.Date(et,tz = "EST"),tz = "EST")
 dt <- as.numeric(difftime(et,st,unit="secs"))
 ev <- sort(runif(N, 0, dt))
 strftime(st+ev, format="%m/%d/%Y")
}
simulateData = function(n = 100, l = 10) {
 data.table(ID = rep(1:n,l), value = rnorm(n*l), date = randDate(n*l))
}
tsData = simulateData()
head(tsData)

##    ID      value       date
## 1:  1 -1.0136803 01/01/2014
## 2:  2  0.5559218 01/01/2014
## 3:  3 -0.1115958 01/02/2014
## 4:  4  0.5620531 01/02/2014
## 5:  5 -0.0299220 01/02/2014
## 6:  6  0.8540974 01/02/2014

nrow(tsData)

## [1] 1000

nrow(limitTimeSeries(tsData, "date", fromDate ="07/01/2014"))

## [1] 517

nrow(limitTimeSeries(tsData, "date", untilDate ="11/01/2014"))

## [1] 828

nrow(limitTimeSeries(tsData, "date", fromDate="07/01/2014", untilDate ="11/01/2014"))

## [1] 345

[Package SparkBeyond version 0.1 Index]