modelBuildingControl {SparkBeyond}	R Documentation

modelBuildingControl

Description

modelBuildingControl

Usage

modelBuildingControl(algorithmsWhiteList = NA, evaluationMetric = NA,
  crossValidation = 5)

Arguments

algorithmsWhiteList:
Optional. A list of strings that represents the set of algorithms to run. NA by default

evaluationMetric:
Optional. A string representing the evaluation metric. Should be either "AUC", "PREC", or "RMSE". NA by default automatically selects AUC for classification and RMSE for regression.

crossValidation:
Optional. Integer value representing how many cross validation splits should be used. 5 by default.

[Package SparkBeyond version 0.1 Index]