offsetTime {SparkBeyond}	R Documentation

offsetTime

Description

Offsets a time-date column by a reference date to create a relative time series with respect to the reference date. Currently work only on data.table

Usage

offsetTime(data, dateCol = "SB_times_col", refDate,
  datesFormat = "%m/%d/%Y", units = "days")

Arguments

data:
data.table to be modified.

dateCol:
The column name in data that will be modified. "SB_times_col" by default

refDate:
The reference date to use.

datesFormat:
the format of the from/until dates. "%m/%d/%Y" by default.

units:
the time span unit to use when creating the reference. Based on difftime definitions. "days" by default.

Value

NA will be returned and the input file will be modified.

[Package SparkBeyond version 0.1 Index]