preProcessingControl {SparkBeyond}	R Documentation

preProcessingContorl

Description

preProcessingContorl

Usage

preProcessingControl(emptyValuePolicy = NA, fileEncoding = NA)

Arguments

emptyValuePolicy:
Controls how empty values in the data are being handled NA by default will replace numeric value with median value, and strings with empty strings.

fileEncoding:
Optional. Options are: "ISO-8859-1", "UTF-8", "US-ASCII". NA by default will try to automatically find the best encoding.

[Package SparkBeyond version 0.1 Index]