reportsList {SparkBeyond}	R Documentation

Shows explantion of all the reports produced by the SparkBeyond enginet

Description

Shows explantion of all the reports produced by the SparkBeyond enginet

Usage

reportsList()

[Package SparkBeyond version 0.1 Index]