writeToServer {SparkBeyond}	R Documentation

writeToServer

Description

A function to write a dateframe to the server. Useful for passing a dataframe to the server for feature search / learning purposes.

Usage

writeToServer(data, filename = NA, prefix = "data_in")

Arguments

data
Data frame or table to export to the server.

filename:
Optional. define a name to save the data to. NA by default.

Value

A filepath to the file on the server that was created.

[Package SparkBeyond version 0.1 Index]