Session-class {SparkBeyond}	R Documentation

SB object that encapsulates a session

Description

SB object that encapsulates a session

Fields

artifact_loc

String location pointing to the model artifact.

modelBuilt

Indication for whether only a feature search was performed or a full model was created.

Methods

buildNumber()

Returns the build number in which the model was generated.

createPackage(sampleData = NULL, createRestAPIpackage = FALSE)

Create a sharable package for the model. sampleData can be used to a sample data to the package and test it. Only first 20 rows of the sample data will be used. createRestAPIpackage is a boolean indicator for whether to create a package for prediction via command line (set to FALSE) or via programmatic REST API call(TRUE).

enrich(data, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. data is a dataframe to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

enrich.file(file, featureCount = NA, writePredictionColumnsOnly = TRUE)

Returns a data frame containing the enrichedData. file is a path to the file to be enriched. Set featureCount in order to limit the number of returned features. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

evaluate()

Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.

features()

Returns a dataset with top feature information

initialize(artifact_loc, modelBuilt = FALSE)

initializes a session using a string provided as loc.

liftFromPrediction(predictionResult, overrideDesiredClass = NA, title = NA, percentOfPopulationToPlot = 0.2)

Returns lift from a created model and generates three plots. predictionResult is a dataframe to be analyzed, overrideDesiredClass the class in the label column to check the lift for (e.g. '1'), title optional: a title for the plot. percentOfPopulationToPlot optional: limit the plot to the top percent of the data (x axis).

predict(data, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. data is a dataframe to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

predict.file(file, writePredictionColumnsOnly = TRUE)

Returns prediction on a created model. file is the path of the file to be predicted. Set writePredictionColumnsOnly to TRUE to return only prediction and probabily columns rather than the entire dataset.

showConfusionMatrix(normalized = FALSE)

Shows a confusion matrix of a model.

showExtractors()

Shows extractors.

showFeatureClusters()

Shows the representative feature clusters pdf report.

showFeaturesTest()

Shows features performance on test.

showFeaturesTrain()

Shows features performance on train.

showFields()

Shows fields.

showFunctions()

Shows functions.

showInputSchema()

Shows the input schema.

showModelComparison()

Shows cross validation of various algorithms tested to create a model.

showROC()

Shows ROC of the model.

showROC_CV()

Shows ROC of cross validation of various algorithms tested to create a model.

showReport(report_name = NA)

report_name should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, train_features, train_unweighted_features, test_unweighted_features,field, function, InputSchema, modelComparison, roc_best, roc_CV

status()

Checking the status of the session.

waitForProcess()

Blocking the R console until session is finished.

Examples

# Learning example
## No test: 
session = learn("titanic", getData("titanic_train"), target="survived",algorithmsWhiteList = list("xgboostClassifier"),  scoreOnTestSet = TRUE, useCachedFeatures=TRUE)

## [1] "Notice: latest build version was not available - if this issue continues please notify SparkBeyond."
## [1] "Training on  ~/sparkBeyond/data_indf7033e47e71.tsv"
## [1] "Calling: http://127.0.0.1:9001/rapi/learn"

## Error in function (type, msg, asError = TRUE) : Failed to connect to 127.0.0.1 port 9001: Connection refused

#session = featureSearch("titanic", getData("titanic_train"), target="survived")
enriched = session$enrich(getData("titanic_train"), featureCount = 10)

## Error in eval(expr, envir, enclos): object 'session' not found

colnames(enriched)

## Error in is.data.frame(x): object 'enriched' not found

predicted = session$predict(getData("titanic_test"))

## Error in eval(expr, envir, enclos): object 'session' not found

colnames(predicted)

## Error in is.data.frame(x): object 'predicted' not found

predicted[1:5,c("survived_predicted", "probability_0", "probability_1")]

## Error in eval(expr, envir, enclos): object 'predicted' not found

eval = session$evaluate()

## Error in eval(expr, envir, enclos): object 'session' not found

#session$showFeatures()
#session$showConfusionMatrix()
## End(No test)

[Package SparkBeyond version 0.1 Index]